+++ package.json
@@ -12,7 +12,7 @@
     "glob": "~3.2.6",
     "optimist": "~0.6.0",
     "mkdirp": "~0.3.5",
-    "imagick": "git+ssh://git@github.com:tjfontaine/node-imagick.git"
+    "handlebars": "~1.0.12"
   },
   "devDependencies": {
     "load-grunt-tasks": "~0.1.0",
